pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "dogesec_commons"
version = "0.0.1"
authors = [
  { name="DOGESEC", email="noreply@dogesec.com" },
]
description = "A package containing views we reuse between various DOGESEC products"
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
]
dependencies = [
    "djangorestframework>=3.15.2",
    "drf-spectacular>=0.27.2",
    "python-arango>=7.9.1",
    "stix2",
]

[tool.hatch.metadata]
allow-direct-references = true

[project.urls]
Homepage = "https://github.com/muchdogesec/stixifier_commons"
Issues = "https://github.com/muchdogesec/stixifier_commons/issues"

[project.optional-dependencies]
stixifier = [
    "txt2stix @ https://github.com/muchdogesec/txt2stix/archive/main.zip",
    "file2txt @ https://github.com/muchdogesec/file2txt/archive/main.zip",
    "stix2arango @ https://github.com/muchdogesec/stix2arango/archive/main.zip",
]
